+++ src/json_codecs.ts
@@ -6,3 +6,4 @@ export {str, num, bool, date, list, map} from './json_codecs/basic';
 export {enumToString} from './json_codecs/enum_to_string';
 export {model} from './json_codecs/model_to_json';
 export {union} from './json_codecs/model_union_to_json';
+export {recordCodec} from './json_codecs/record_to_json';