+++ test/src/json_codecs.spec.ts
@@ -3,6 +3,7 @@ import {enumToStringTests} from './json_codecs/enum_to_string.spec';
 import {modelPropertyToJsonTests} from './json_codecs/model_property_to_json.spec';
 import {modelToJsonTests} from './json_codecs/model_to_json.spec';
 import {modelUnionToJsonTests} from './json_codecs/model_union_to_json.spec';
+import {recordToJsonTests} from './json_codecs/record_to_json.spec';
 import {stringCaseConverterTests} from './json_codecs/string_case_converters.spec';
 import {objectToJsonTests} from "./json_codecs/object_to_json.spec";
 
@@ -14,6 +15,7 @@ export function jsonCodecsTests() {
         modelPropertyToJsonTests();
         modelToJsonTests();
         modelUnionToJsonTests();
+        recordToJsonTests();
         stringCaseConverterTests();
     });
 }